updated types of ui components
