+++ packages/overlay/src/types.ts
@@ -1,4 +1,3 @@
-import { ComponentPropsWithoutRef } from 'react';
 import { type Integration } from './integrations/integration';
 
 export type ExperimentName = 'sentry:focus-local-events';
@@ -117,19 +116,4 @@ export type WindowWithSpotlight = Window & {
   __spotlight?: {
     eventTarget?: EventTarget;
   };
-};
-
-export type SidePanelProps = Omit<ComponentPropsWithoutRef<'div'>, 'className'> & {
-  backto: string;
-};
-
-export type CrumbProps = {
-  id: string;
-  label: string;
-  link?: boolean;
-  to?: string;
-};
-
-export type BreadcrumbProps = {
-  crumbs: CrumbProps[];
 };
