+++ packages/overlay/src/ui/Breadcrumbs/index.ts
@@ -1,1 +1,2 @@
 export { default } from './Breadcrumbs';
+export type { BreadcrumbProps, CrumbProps } from './Breadcrumbs';
