+++ packages/overlay/src/ui/Chart/RingChart/index.ts
@@ -1,1 +1,2 @@
 export { default } from './RingChart';
+export type { RingChartProps } from './RingChart';
