+++ packages/overlay/src/ui/SidePanel/index.ts
@@ -1,1 +1,2 @@
 export { SidePanelHeader, default } from './SidePanel';
+export type { SidePanelProps } from './SidePanel';
