+++ packages/overlay/src/ui/Tag/Tag.tsx
@@ -1,4 +1,6 @@
-export default function Tag({ tagKey, value }: { tagKey: string; value: string }) {
+export type TagProps = { tagKey: string; value: string };
+
+export default function Tag({ tagKey, value }: TagProps) {
   return (
     <div className="border-primary-300 bg-primary-900 divide-x-primary-300 inline-flex divide-x whitespace-nowrap rounded-full border font-mono text-sm">
       <div className="px-2 py-1 font-semibold">{tagKey}</div>
