+++ packages/overlay/src/ui/Tag/index.ts
@@ -1,1 +1,2 @@
 export { default } from './Tag';
+export type { TagProps } from './Tag';
