+++ packages/overlay/src/ui/Tooltip/index.ts
@@ -1,1 +1,2 @@
 export { default } from './Tooltip';
+export type { TooltipProps } from './Tooltip';
